bug fix: block id are truncated to 32bit

+++ toolchain/udbfslib/allocate_block_id.c
@@ -1,4 +1,4 @@
-// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/allocate_block_id.c,v 1.1 2003/10/11 13:14:19 bitglue Exp $
+// $Header: /home/xubuntu/berlios_backup/github/tmp-cvs/uuu/Repository/toolchain/udbfslib/allocate_block_id.c,v 1.2 2003/10/12 18:14:07 instinc Exp $
 
 #define _LARGEFILE_SOURCE
 #define _LARGEFILE64_SOURCE
@@ -14,7 +14,7 @@
 uint64_t	udbfs_allocate_block_id(
     UDBFSLIB_MOUNT	*mount ) {
 
-  return( (uint32_t)udbfslib_allocate_bit(
+  return( udbfslib_allocate_bit(
 	mount->block_bitmap,
 	mount->block_bitmap_size,
 	&mount->free_block_count) );
